pequeños cambios al front y al back

+++ database/models/Product.js
@@ -13,7 +13,7 @@ module.exports = (sequelize, dataTypes) => {
             allowNull: false
         },
         price: {
-            type: dataTypes.DECIMAL(10),
+            type: dataTypes.INTEGER,
             allowNull: false
         },
         discount: {
